NOISSUE, trivial code change.

+++ addons/member_communication/member_communication.addon.php
@@ -32,10 +32,10 @@ if($this->module != 'member' && $called_position == 'before_module_init')
 	$oMemberController->addMemberMenu('dispCommunicationFriend', 'cmd_view_friend');
 	$oMemberController->addMemberMenu('dispCommunicationMessages', 'cmd_view_message_box');
 
-	if($addon_info->use_alarm != 'N');
+	$flag_file = _XE_PATH_ . 'files/member_extra_info/new_message_flags/' . getNumberingPath($logged_info->member_srl) . $logged_info->member_srl;
+	if($addon_info->use_alarm != 'N' && file_exists($flag_file));
 	{
 		// Pop-up to display messages if a flag on new message is set
-		$flag_file = _XE_PATH_ . 'files/member_extra_info/new_message_flags/' . getNumberingPath($logged_info->member_srl) . $logged_info->member_srl;
 		$new_message_count = (int) trim(FileHandler::readFile($flag_file));
 		FileHandler::removeFile($flag_file);
 		Context::loadLang(_XE_PATH_ . 'addons/member_communication/lang');
